+++ taxi/commands/commit.py
@@ -28,12 +28,9 @@ from .types import DateRange
 @click.pass_context
 def commit(ctx, f, force_yes, date, not_today):
     """
-    Usage: commit
-
     Commits your work to the server. The [date] option can be used either as a
     single date (eg. 20.01.2014), as a range (eg. 20.01.2014-22.01.2014), or as
     a range with one of the dates omitted (eg. -22.01.2014).
-
     """
     timesheet_collection = get_timesheet_collection_for_context(ctx, f)
 
@@ -61,7 +58,9 @@ def commit(ctx, f, force_yes, date, not_today):
     try:
         # Push entries
         for timesheet in timesheet_collection.timesheets:
-            entries_to_push = get_entries_to_push(timesheet, date)
+            entries_to_push = get_entries_to_push(
+                timesheet, date, ctx.obj['settings']['regroup_entries']
+            )
 
             for (entries_date, entries) in entries_to_push.items():
                 for entry in entries:
@@ -117,13 +116,7 @@ def backends_post_push(backends_entries):
 
 def comment_timesheets_entries(timesheet_collection, date):
     for timesheet in timesheet_collection.timesheets:
-        # Mark local entries as pushed without pushing them
-        local_entries = timesheet.get_local_entries(date)
         pushed_entries = get_entries_to_push(timesheet, date)
-
-        for (entries_date, entries) in six.iteritems(local_entries):
-            for entry in entries:
-                entry.commented = True
 
         for (entries_date, entries) in pushed_entries.items():
             for entry in entries:
@@ -142,8 +135,7 @@ def comment_timesheets_entries(timesheet_collection, date):
         timesheet.file.write(timesheet.entries)
 
 
-def get_entries_to_push(timesheet, date):
+def get_entries_to_push(timesheet, date, regroup=True):
     return timesheet.get_entries(
-        date, exclude_ignored=True,
-        exclude_local=True, exclude_unmapped=True, regroup=True
+        date, exclude_ignored=True, exclude_unmapped=True, regroup=regroup
     )