+++ taxi/commands/show.py
@@ -2,30 +2,78 @@ from __future__ import unicode_literals
 
 import click
 
+from taxi.aliases import aliases_database
+
 from .base import cli
 
 
-@cli.command(short_help="Show the details and activities of a project.")
-@click.argument('project_id', type=int)
-@click.argument('backend', required=False)
+@cli.command(short_help="Resolve any object passed to it (aliases, projects, "
+                        "etc).")
+@click.argument('search', nargs=1)
 @click.pass_context
-def show(ctx, project_id, backend):
+def show(ctx, search):
     """
-    Usage: show project_id [backend]
+    Resolves any object passed to it (aliases, projects, etc).
 
-    Shows the details of the given project_id (you can find it with the search
-    command).
+    This will resolve the following:
 
+    \b
+        - Aliases (my_alias)
+        - Mappings (123/456)
+        - Project ids (123)
     """
+    matches = {'aliases': [], 'mappings': [], 'projects': []}
+    projects_db = ctx.obj['projects_db']
+
+    matches = get_alias_matches(search, matches)
+    matches = get_mapping_matches(search, matches, projects_db)
+    matches = get_project_matches(search, matches, projects_db)
+
+    ctx.obj['view'].show_command_results(search, matches, projects_db)
+
+
+def get_alias_matches(search, matches):
+    if search in aliases_database:
+        matches['aliases'].append(aliases_database[search])
+
+    return matches
+
+
+def get_mapping_matches(search, matches, projects_db):
+    if '/' not in search:
+        return matches
+
+    mapping = search.split('/', 1)
     try:
-        project = ctx.obj['projects_db'].get(project_id, backend)
-    except IOError:
-        raise Exception("Error: the projects database file doesn't exist. "
-                        "Please run `taxi update` to create it")
+        mapping = tuple(int(m) for m in mapping)
+    except ValueError:
+        return matches
 
-    if project is None:
-        ctx.obj['view'].err(
-            "Could not find project `%s`" % (project_id)
-        )
-    else:
-        ctx.obj['view'].project_with_activities(project)
+    alias_mappings = aliases_database.filter_from_mapping(mapping)
+    for alias, alias_mapping in alias_mappings.items():
+        matches['mappings'].append((alias_mapping, alias))
+
+    project = projects_db.get(mapping[0])
+    project_activity = None
+
+    for activity in project.activities:
+        if activity.id == mapping[1]:
+            project_activity = activity
+
+    if project and project_activity:
+        matches['projects'].append((project, project_activity))
+
+    return matches
+
+
+def get_project_matches(search, matches, projects_db):
+    try:
+        project_id = int(search)
+    except ValueError:
+        return matches
+
+    project = projects_db.get(project_id)
+    if project:
+        matches['projects'].append((project, None))
+
+    return matches